requirements.txt
Flask>=2.2.2
boto3==1.37.6 
Pillow
image